pyproject.toml
[tool.poetry]
name = "telegram-notes-bot"
version = "0.1.0"
description = "Telegram bot for saving text messages to markdown notes"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.0"
python-dotenv = "^1.0.0"
grpcio = "^1.62"
protobuf = ">=5.0"
psycopg2-binary = "^2.9"

[tool.poetry.group.dev.dependencies]
grpcio-tools = "^1.62"
ruff = "^0.14.4"
pre-commit = "^4.4.0"
pytest = "^8.0"
pytest-cov = "^6.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=core --cov=notifications --cov-report=term-missing --cov-report=html:htmlcov"

[tool.coverage.run]
source = ["core", "notifications"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"